chore: precompile shaders
Removes the need for all developers to have
slangc installed. Our shaders are pretty stable,
so I think overall this is an improvement.

All shader formats we used are though of an
interchange format, so the use case we have
here is totally fine. The graphic driver will
compile / optimize the shaders anyhow.
